pyproject.toml
[project]
name = "gcp-doc-ai-showcase"
version = "0.1.0"
description = "Interactive Streamlit showcase for GCP Document AI with visual bounding box overlays"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Andrey Vykhodtsev" }]
dependencies = [
    "streamlit>=1.28.0",
    "google-cloud-documentai>=2.20.0",
    "google-auth>=2.22.0",
    "google-api-core>=2.11.0",
    "protobuf>=4.23.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.3",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "typing-extensions>=4.5.0",
    "pandas>=2.0.0",
]
